requirements.txt
# AI On-Call Agent Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
requests>=2.31.0
redis>=5.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.13.0
celery>=5.3.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.0
aiofiles>=23.2.1
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.1.0
openai>=1.3.0
langchain>=0.1.0
transformers>=4.35.0
typer>=0.9.0
rich>=13.7.0
click>=8.1.0
markdown2>=2.4.0
weasyprint>=60.0
reportlab>=4.0.0

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0
pre-commit>=3.5.0
jupyter>=1.0.0
ipython>=8.17.0
